feat: Moving checkIfDryRunWasRequested ond findPackageJson out to functions dir
