+++ test/checkIfDryRunWasRequested.test.js
@@ -1,16 +1,20 @@
 const { describe, it } = require("node:test");
 const assert = require("node:assert");
 
-const { checkIfDryRunWasRequested } = require("../index.js");
+const checkIfDryRunWasRequested = require('../functions/checkIfDryRunWasRequested').default;
 
 describe("Test suite: checkIfDryRunWasRequested", () => {
+  const context = {
+    dryRunParameter: 'dryRun'
+  }
+
   it('Should return true when among the arguments we pass --dryRun', () => {
     const params = [
       '/Users/user/.nvm/versions/node/v18.12.1/bin/node',
       '/Users/user/Projects/kelsus/api-spot-package/bin/npx.js',
       '--dryRun'
     ];
-    const dryRunWasRequested = checkIfDryRunWasRequested(params);
+    const dryRunWasRequested = checkIfDryRunWasRequested(context, params);
 
     assert.equal(dryRunWasRequested, true);
   });
@@ -22,7 +26,7 @@ describe("Test suite: checkIfDryRunWasRequested", () => {
       '--dryRun=',
       '--serviceUrl=https://some.url.com'
     ];
-    const dryRunWasRequested = checkIfDryRunWasRequested(params);
+    const dryRunWasRequested = checkIfDryRunWasRequested(context, params);
 
     assert.equal(dryRunWasRequested, true);
   });
@@ -34,7 +38,7 @@ describe("Test suite: checkIfDryRunWasRequested", () => {
       '--serviceUrl=https://some.url.com',
       '--dryRun=true'
     ];
-    const dryRunWasRequested = checkIfDryRunWasRequested(params);
+    const dryRunWasRequested = checkIfDryRunWasRequested(context, params);
 
     assert.equal(dryRunWasRequested, true);
   });
@@ -46,7 +50,7 @@ describe("Test suite: checkIfDryRunWasRequested", () => {
       '--serviceUrl=https://some.url.com',
       '--dry-run=true'
     ];
-    const dryRunWasRequested = checkIfDryRunWasRequested(params);
+    const dryRunWasRequested = checkIfDryRunWasRequested(context, params);
 
     assert.equal(dryRunWasRequested, false);
   });
@@ -57,7 +61,7 @@ describe("Test suite: checkIfDryRunWasRequested", () => {
       '/Users/user/Projects/kelsus/api-spot-package/bin/npx.js',
       '--serviceUrl=https://some.url.com'
     ];
-    const dryRunWasRequested = checkIfDryRunWasRequested(params);
+    const dryRunWasRequested = checkIfDryRunWasRequested(context, params);
 
     assert.equal(dryRunWasRequested, false);
   });
